Refactor: Catalogue Tabs and Remove Unused Components
