+++ app/Enums/UI/Catalogue/IrisProductTabsEnum.php
@@ -11,7 +11,6 @@ enum IrisProductTabsEnum: string
     use EnumHelperTrait;
     use HasTabsWithQuantity;
 
-    case OVERVIEW = 'overview';
     case PRODUCTS = 'products';
 
     public function blueprint(ProductCategory $parent): array
@@ -19,10 +18,6 @@ enum IrisProductTabsEnum: string
         $products = $parent->stats->number_products_state_active + $parent->stats->number_products_state_discontinuing;
 
         return match ($this) {
-            IrisProductTabsEnum::OVERVIEW => [
-                'title' => __('Overview'),
-                'icon'  => 'fal fa-tachometer-alt-fast',
-            ],
             IrisProductTabsEnum::PRODUCTS => [
                 'title' => __('Products') . " ({$products})",
                 'icon'  => 'fal fa-cube',
